+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * `SelectList` inherits parent width more naturally, and no longer requires an explicit numeric width be set
 * `SelectList` supports `inverted` prop, to expand the selection options upwards rather than downwards
 * `Tag` enables transitions, for consistency with other components
-* Support ref forwarding to the underlying native DOM element for `Button`, `TextField`
+* Support ref forwarding to the underlying native DOM element for `Button`, `TextArea`, `TextField`
 * Standardize transitions across all components with custom easing function
 
 ## v1.2.1